fix: OP_ABORT Int→Byte conversion in buildAbortPacket

+++ app/src/main/java/com/btsec/testtool/data/bredr/ObexClient.kt
@@ -371,7 +371,7 @@ class ObexClient(
     }
 
     private fun buildAbortPacket(): ByteArray {
-        return byteArrayOf(OP_ABORT, 0x00, 0x03)
+        return byteArrayOf(OP_ABORT.toByte(), 0x00, 0x03)
     }
 
     // ── Header writers ──
